requirements.txt
fastapi
pandas
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
pydantic
streamlit